requirements.txt
python-dotenv==1.0.0
pydantic==2.4.2
pytest~=7.4.2
httpx~=0.25.0
django==4.2.5
django-environ==0.11.2
psycopg2-binary==2.9.8
django-extensions==3.2.3
django-fsm==2.8.1
gunicorn==21.2.0
whitenoise==6.5.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
pydot==1.4.2
graphviz==0.20.1
Faker===19.10.0
redis==5.0.1
celery==5.3.4